PNG fixes, starting to look good
git-svn-id: svn://svn.freevo.org/freevo/trunk/freevo@145 cc3e1ea1-1e01-0410-8d68-8b121e83a9d5

+++ osd_server/osd_fb/readpng.h
@@ -11,7 +11,7 @@
 #endif
 
 /* Returns OK or ERROR */
-extern int read_png (char *file_name, uint32 **ppBitmap,
+extern int read_png (char *file_name, uint8 **ppBitmap,
                      uint16 *pWidth, uint16 *pHeight);
 
 #endif /* _INC_READPNG_H_ */
